refactor with common
