+++ __tests__/01_basic_spec.tsx
@@ -1,6 +1,6 @@
 import {
-  atomWithTanstackMutation,
   atomsWithTanstackInfiniteQuery,
+  atomsWithTanstackMutation,
   atomsWithTanstackQuery,
   queryClientAtom,
 } from '../src/index'
@@ -10,6 +10,6 @@ describe('basic spec', () => {
     expect(queryClientAtom).toBeDefined()
     expect(atomsWithTanstackQuery).toBeDefined()
     expect(atomsWithTanstackInfiniteQuery).toBeDefined()
-    expect(atomWithTanstackMutation).toBeDefined()
+    expect(atomsWithTanstackMutation).toBeDefined()
   })
 })
