+++ src/index.ts
@@ -1,4 +1,4 @@
 export { queryClientAtom } from './queryClientAtom'
 export { atomsWithTanstackQuery } from './atomsWithTanstackQuery'
 export { atomsWithTanstackInfiniteQuery } from './atomsWithTanstackInfiniteQuery'
-export { atomWithTanstackMutation } from './atomWithTanstackMutation'
+export { atomsWithTanstackMutation } from './atomsWithTanstackMutation'
